Engine: Lock framerate to 60FPS
This might come back to bite me in the arse, but let's do it for now, as
it removes a lot of the nasty `dt` shit from the game loop. I doubt we
really need to worry about not hitting the frame-rate for a game like
Mario anyway (not to mention the physics sim is very simple).