Gfx: Make `Texture` moveable
The texture class is now movable, meaning one texture can be moved to
another one, invalidating the original. This is especially handy with
things such as a resource loader class. This will need to return a
local class object, and as such (due to `GFX::Texture` containing a
`std::unique_ptr<>`, requires a move constructor to exist, as it
implicitly deleted.